We are officially Uncooked
3:05

after ensuring the rest of merch works, i can proudly say that I made this. I spend a total of 125:25... yes, 125 1/2 hours of coding at home on this project XD.

Though I dont plan on going into this field (because the competition is far too high for me) it was fun. and with that... this is my last commit...

Adios 👋